+++ lib/src/utils/extensions.dart
@@ -77,8 +77,8 @@ extension BuildContextEasyLocalizationExtension on BuildContext {
 
   /// Sets new translations and saves it to storage if translationsStorage
   /// is not null
-  Future<void> replaceTranslations(
+  Future<void> injectTranslations(
     List<Map<LocalizationKey, LocalizationValue>> next,
   ) =>
-      ImLocalizedApp.of(this)!.replaceTranslations(next);
+      ImLocalizedApp.of(this)!.injectTranslations(next);
 }